requirements.txt
fastapi>=0.110.0,<1.0.0
openai>=1.40.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
pypdf>=4.0.0,<6.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
